+++ docs/ChangeLog.rst
@@ -5,6 +5,14 @@ Release Notes
 Version 1.0 (in development)
 ============================
 
+Main changes:
+-------------
+
+* Added ability to submit a form without updating `StatefulBrowser` internal
+  state. This means you get a response from the form submission, but your
+  browser "stays" on the same page. Useful for handling forms that result in
+  a download of a file or opening a new window.
+
 Bug fixes
 ---------